Improve MWOAuthSessionProvider authorization test
If a non-owner could somehow approve an owner-only consumer, the
existing code would allow that consumer to be used by the non-owner even
though owner-only consumers are supposed to be only usable by the owner.

That isn't possible since MWOAuthServer::authorize() refuses to work on
owner-only consumers, but we should fix the check here anyway.

Change-Id: Iaa0c99ee57449ac296b09d082cd8f0e4cb5fcd90