Replace deprecated OutputPage::parseInline()
The OutputPage::parseInline() method emits untidy output and
is often used with the wrong user interface/content language
selection.  Replace with new method added in 1.33 which is
tidy and consistent.

In addition, replace Message::text() with Message::plain() calls to
avoid double-expanding `{{...}}` constructs when the result of
Message::text() is ultimately going to be passed to the wikitext
parser.

Bug: T198214
Bug: T206574
Depends-On: Ica79c2acbc542ef37f971c0be2582ae771a23bd0
Change-Id: I4029ab115ccc0e518928e54505f1aa2a4631635a