Respect language code in onMessagesPreLoad hook
Give the ConsumerAccessControl object a DerivativeContext with
the correct language, so that it will use that language when calling
$this->msg() internally. Previously it was using the main context and
ignoring the requested language.

This also means that this hook will never attempt to load the current
user in order to discover which language to use, which means it can
safely be called from entry points where sessions are disabled (like
load.php).

Bug: T201574
Change-Id: I58901657daf3e82229a303d8fe7a2c051ffe7634