Add "OAuth CID: $consumerId" tag to patrol log
When an existing RecentChange is marked as patrolled, the save() method
is not used, and so our existing hook doesn’t get a chance to add the
tag for the OAuth consumer if the request was made via OAuth. Subscribe
to the MarkPatrolled hook, which recently gained a &$tags argument, to
add the tag here instead.

Bug: T219655
Change-Id: I2c929f8ef6b5da54bfd75e185f409602470fddd7
Depends-On: Ie5e6f820bbf399ec639e715afd908f78bf5c8e9a