build: Add mediawiki/mediawiki-phan-config
Removed @method, because phan does not understand them and it seems
there are not needed, because the return type override with this
annoation is not needed due to use of static in the parent
implementation. Also parent is final, which disallow overrides.
The only benefit is, that the first argument is the correct type and not
*any* subclass of the needed class. But phan does not want the types
without typehinting in the parent method

Bug: T224761
Change-Id: I5f60fac2db3edaec0ca3a5b50deadf97353f5c8e