Avoid mentioning real names on wikis which do not use it
Do not ask for the user's permission to reveal their real name
when the wiki is not actually configured to use real names.
Since auth-only grants either give access to both the email and
the real name or neither, this was a common source of user angst.

Also make message documentation for grants options a little saner.

Bug: T151515
Change-Id: Ie2aa8d4b019c6886e785f90d1f4455b225c3bd0b