Add alias for old namespaced version of MWOAuthToken
Needed for WMF production whilst we are running two versions of MW
which use the same cache.

Shouldn't be needed when .30 is everywhere and stable and therefore not
need carrying forward to next weeks train.

Cache TTL is only 10 minutes, but the multiple MW versions make this
a bit more complex

Bug: T251306
Change-Id: I32f397272e6ec65972f0e1a772db55e1bc76ae98