AuthorizationProvider: don't use expiration dates that are so large that they break PHP 7.3.23+
Reduce "infinite" expiry from 292 billion years to 1000 years. That's
still long enough that I won't be around to deal with the bug reports
that will inevitably come in due to extremely persistent OAuth clients
needing to reauthenticate, but short enough that it doesn't overflow
64-bit integers used in intermediate calculations in PHP's date
library, and short enough to fit in a 4-digit year, thus avoiding an
exception from ConvertibleTimestamp.

Bug: T269502
Change-Id: I80461d2fcf02b41581e29e73f5661b667158e84e