Handle non-existent local user in ScopeRepository::finalizeScopes
I can think of three ways this could happen:
* incorrect central user ID sent by client
* user is globally locked
* user does not have a local account (either autocreation did not
  happen yet or it failed due to permission issues, AbuseFilter
  rules, the local account is already taken by someone else etc)

In the first two cases, the request need to be rejected (ideally
with better logging but that's true of many places in the extension).
In the third, ideally we'd try to autocreate (see e.g. T208443), but
for now let's just also fail.

Bug: T294015
Change-Id: I1b1040fb0f0f4d33f299a77aaba6449fa13438c6