Add AbuseFilter variable for used OAuth consumer
To help some wikis help dealing with abuse coming from certain
OAuth-backed tools, set the oauth_consumer AbuseFilter variable to the
ID of the used OAuth consumer, if any. This information is already
public in the form of edit tags.

This is technically done by implementing the relevant AbuseFilter hooks,
and setting the variables based on data stored in the OAuth session
provider. It relies on the User::getSession() method, which I think is
file (at least it works on my machine). At least I'm not aware of any
better ways to implement that;

Note that this variable will only be set if the consumer is not set as
owner-only. This is what we want, since owner-only consumers are not
exposed as tags either and malicious actors can easily get more
owner-only consumers while changing the consumers specific public tools
use is not as trivial.

Bug: T298281
Change-Id: Ia7c5200d785788f329d328eb8da21afb25dbc5ff