maint/createOAuthConsumer.php: Specify oauth2GrantTypes
As of MW 1.39 when using this maintenance script you'll
get an error that looks something like this

Missing value for "oauth2GrantTypes" field

I have taken the default value from the special page and
added it to the script to make it work again.

This was found while making the 1.39 compatible wikibase
release pipeline release.

Bug: T329306
Change-Id: Id2002cb7129528562d4511c7694edd9f7e49672c