Fix accessibility in OAuth authorization dialog
We need to reattach the form in a different place in the dialog to
avoid accessibility issues, because the OOUI message dialog isn't
really intended for putting an entire form into it.

When doing it, we must wait until the dialog has opened rather than
only until it starts opening, otherwise the dialog's opening process
moves the content back to the "wrong" place.

Bug: T329796
Change-Id: Ie1b42f8b7194d9e3b00736ffa099cef8b91d5e60