Allow management form validators to return a status
The SubmitControl validators used in the consumer proposal, acceptance,
authorizarion etc. forms could only return true or false, which is not
terribly user-friendly. Allow returning a StatusValue so that the
message can refelect the specific way in which the value is wrong, and
some problems can be marked as optional using StatusValue's warnings
system. Optional problems can be ignored by cheching an extra box.
(This will be used for things that are usually mistakes in practice
but in theory don't have to be, e.g. using a callback URL that only
consists of an origin.)

There is probably a reasonable method for only showing an "ignore
warnings" checkbox on the form when there are warnings; I haven't
found it though, so the patch uses a CSS hack to hide the checkbox
when it's not relevant.

Change-Id: I2d71ff12c6dc4cf0a697f081bc93976fcc041768