Hide new consumer notification if it was handled by another admin
Do not show notifications about proposed new consumers if they
aren't in the 'proposed' stage anymore (ie. another OAuth admin
has approved or rejected them).

Originally this wasn't done to allow admins to review actions by
other admins, but that wasn't particularly useful (and if wanted,
it could still be done via Special:Log), and with the current
volume of new consumers it became highly disruptive.

This has the disadvantage that even if you are the reviewing admin,
the notification disappears once you act on it, but that seems
acceptable (and not worth doing an extra query to check if the
current user is the one who reviewed the consumer).

Change-Id: I4b659df117b204e2769e892c6d0a883276f5d511