Auto-approve apps with authentication-only grants
There is very little risk on public wikis in approving OAuth apps
which use one of the "User identity verification only, no ability to
read pages or act on a user's behalf.", "User identity verification
only with access to real name and email address, no ability to read
pages or act on a user's behalf.", or "Basic rights" grants, and in
practice such requests tend to be approved unconditionally. They also
make up for a large part of OAuth app requests, and recently the
request volume for Wikimedia apps increased to the point where review
is a significant burden for OAuth admins.

To improve that, the patch:
* adds a Workflow class for centralizing essential business logic
  related to the app review workflow;
* adds a Workflow::consumerCanBeAutoApproved() method which can
  determine that an app is safe enough to not require manual
  approval, and calls it at the appropriate places;
* indicates to the user in the app submission success message wether
  they need to wait for approval;
* introduces a $wgOAuthAutoApprove configuration variable to define
  which grants are compatible with auto-approval (structured in a
  generic way so in the future rules other than grants can be added)

The configuration for Wikimedia wikis is in I3f6904358cc7.

Other changes:
* Introduce a service container.
* Split the confirmation message after submitting a new consumer
  into OAuth 1 and OAuth 2 versions, as these use different
  terminology.

Bug: T67750
Change-Id: I9fd67f554cae1c3bd61b62f0afa5f0b56c6fd0af