Fix default refresh token expiry
It was set to PT1M (1 minute) which doesn't make any sense for a
token that's supposed to be long-lived. Presumably it's a typo and
P1M (1 month) was meant.

Also spell out the intervals in the variable description since most
people probably aren't familiar with the ISO duration notation;
and get rid of the pointless has() checks (it's a variable defined
in extension.json, it can't be missing unless the entire config
framework gets seriously broken).

Also add note that refresh tokens are stored in cache, according to
$wgMWOAuthSessionCacheType, so this change doesn't necessarily
result in refresh tokens being valid for a month.

Bug: T340921
Change-Id: Ie2acc79dee4b755dd5e777c1b83be24a76336e2a