Temporarily allow OAuth on non-API entry points again
This effectively reverts change I81ae25d6a6 (commit e56c316a7b), though
it keeps the improved makeException() and the i18n messages. We need to
allow at least Special:OAuth to execute when OAuth parameters are given
in the request; for now, restore the old behavior of returning null for
all non-API requests, regardless of title. We can limit this to
Special:OAuth (and potentially other pages that need it?) later.

Bug: T341656
Change-Id: Ibbbc4cc59fbf81e65515cab87e15a4714935a8ab