Add recenchanges link to consumer detail view
Add a link to recent changes filtered by the consumer's change tag
to the bottom of the detail view at Special:OAuthListConsumers.
The link is already present in the list view but was accidentally
omitted from the detail view.

It would probably make more sense to only add it for non-owner-only
consumers, but the list view also adds it unconditionally so opting
for consistency for now.

Bug: T224749
Change-Id: I7305190a828e934e52f2022bb379897eba626e93