SpecialMWOAuth: Support &display=popup
Setting this URL parameter will replace the default skin with a
"micro-skin" that omits most of the usual interface elements,
making the page suitable to be displayed in a small popup window.

The same behavior has been implemented for the login/signup forms in
MediaWiki core. See the dependency for details of how it works.

Depends-On: Ic7cbbe98344b25d2e965750e0c4429ce157163ed
Bug: T71246
Change-Id: I190686149e0de8cd4995359c08d267751a4d81f2