Simplify grant factory error handling
The grant parameter is checked by ParamValidator, if despite that
the factory is not found, that's a bug, not a user error. Throw
a LogicError instead of an invalid_request error which is used
in many places in the code which are actually reachable, so
having this unreachable instance just complicates debugging those.

Bug: T369156
Change-Id: I709df5d8047d8d90e8c535379aa2fc880b7579ee