Fix format of POST data in tests
The data in these arrays must be in the same format as the $_POST
array in PHP, and contain only strings, not parsed structures.
New validation added in MediaWiki core in change
I2d00d320f59652040ff914005aad58bf3e724689 caused the tests to fail.

I struggled to figure out how booleans and arrays are actually
represented in REST API POST requests. I'm not sure if this is
correct, but the tests pass.

Change-Id: I0d9457df2fd4de99c1056580be31aa565c1257ec