Hooks::getUsedConsumerTags: Remove second query
As far as I can tell, querying change_tag should be unnecessary:
change_tag_def will already only contain tags that are used on the wiki,
so we know whether the tag is used or not based on whether we can find
an ID for it. So we can completely remove the query which is sometimes
slow on MariaDB 10.5 (shipped with Debian Bullseye; fixed in 10.6).

Bug: T370578
Change-Id: I186b75a434781fd4656f714394c6e4c09ae40a6f
(cherry picked from commit 59192c83e4d57368fcd1ca358d48efc19257ef88)