createOAuthConsumer: Add --oauthVersion option
This makes it possible to create OAuth 2 consumers (clients).

Also fix the definition of --callbackIsPrefix: the option has no
argument and is interpreted as a boolean by its presence or absence
($this->hasOption( 'callbackIsPrefix' )), so it makes no sense to
require it.

The --oauth2IsNotConfidential option isn’t great, but I don’t think
there’s a really great way to represent boolean options that should
default to true (see also discussion at [1]). If reviewers prefer, I’m
open to changing it to --oauth2IsConfidential=true|false or something
else.

[1]: https://wikis.world/@LucasWerkmeister/113772261393987942

Bug: T211568
Bug: T374562
Change-Id: I99c89b2cfada8469276eafc3f7e754e8047efb7c