SpecialPages: Remove unused lang param in MWOAuthConsumerRegistration
Introduced in 195fc8414c9c44 (I999d1773f0df43d015799) and later during
the refactoring that happened in 34947cf7b8485 (Iee3b50794a169b86c44e)
the usage of the lang object parameter was removed.

This patch is a minor cleanup to remove the unused parameter from the
the method signature.

Change-Id: Ide533283ea1214d83a1d503ca294d15648f27bb9