Use central wiki as access token issuer
Use the canonical server URL of the central wiki configured via
$wgMWOAuthCentralWiki as the OAuth 2 access token's 'iss' field.
For CentralAuth wikis, this matches how the issuer field is used
for session cookie JWTs.

Since we don't validate issuers whatsoever, this is a backwards-
compatible change as long as nothing external relies on it. Since
previously the issuer was selected in an arbitrary way (e.g.
whichever wiki you happened to exchange a refresh token on became
the issuer of the access token), it's unlikely anything would.

Bug: T399199
Change-Id: Iee68246a995c6e0b7240f5db8b4fcc005c8c48ba