Deprecate $wgMWOAuthSharedUserIDs
$wgMWOAuthSharedUserIDs=false is the same as
$wgMWOAuthSharedUserSource='local', it's a pointless setting.

Deprecating it is a bit tricky, since a change to the shared user
source would completely break a wiki; we want to keep the default
behavior to do the sensible thing and e.g. use CentralAuth central
identities on CentralAuth-based wikifarms, but also don't want
the behavior to change without warning, when someone intentionally
set $wgMWOAuthSharedUserIDs to false to override default behavior.
So the best we can do is trigger warnings for a while before making
the change.

Also remove references to old hooks that were removed in I622923f4581ae,
and change UserMergeFromTo handling when the extension uses central
IDs (probably won't work, but no worse than not doing anything).

Also fixes some bugs in the $wgMWOAuthSharedUserIDs=false case:
* Util::getCentralUserNameFromId() and Util::getLocalUserFromCentralId()
  used User::newFromId() without validating the ID, and so didn't
  return false when the user was not found
* Utill::getCentralIdFromLocalUser() did not check for user existence
  and likewise didn't return false when it should

Change-Id: Ib4ac7ea6507cb1783b89490f718ed18155d8e6d5