Use GetSessionJwtData hook for OAuth 2 access token JWT
In addition to the existing OAuthClaimStoreGetClaims hook,
also call the GetSessionJwtData hook when generating access
tokens. This makes them more similar to JWTs generated from
SessionManager::getJwtData() (which we can't easily use here
since the JWT handling logic is so fragmented between the
extension and the third-party OAuth 2 library).

Make sure that in case of a conflict OAuthClaimStoreGetClaims
overrides GetSessionJwtData, as it's the one with the more
specific purpose.

Bug: T399199
Change-Id: I9b7ea25f91deada2a7ea4dc06e0fad9897645604