Issue SessionManager-style 'sub' claims for non-owner-only
Issue SessionManager-style 'sub' claims (where the central user ID
is prefixed by 'mw:' and the central ID provider name) for
non-owner-only access tokens.

Follows up If671839c558e which made token validation accept such
tokens.

Owner-only tokens are not changed - they are not really useful for
traffic control which was the motivation for this change, and they
are valid forever so we don't want to introduce a new type of them
lightly.

Bug: T399199
Change-Id: I24112f80fd554cec6f28d135448620b2ace84e2f