Emit deprecation warnings when virtual domain isn't set
Begin notifying third-parties to set virtual domain now that
we support it. On a wiki-farm setup, users can point OAuth to
the shared DB they want the extension to interact with. Note
that setting `db` to `false` will default to the wiki that
made the request (local wiki) otherwise this will be a shared
wiki.

This is actually a followup to I2b5ecfb4b9cf10fdd2e57ce57.

Bug: T409184
Change-Id: I30ea58a8b8b80e6d8a3efc2689a3a3f770581557