Indicate owner-only status to OAuth 2 access token claims hooks
Normal access tokens expire quickly, owner-only access tokens
never expire, which means they are likely to be used differently.
Let the hooks which add claims to these tokens know when a token
is owner-only.

This is done in a very hacky way for now, to avoid having to change
hook signatures. If we decide we need this capability, we might want
to replace the hint mechanism with something nicer.

Bug: T399632
Change-Id: I1802b6975392f333d668fecec5a2afb124e4e44c