Backend: Removed unused IConnectionProvider property
Introduced in I1ad23dadd5eda50d00 (699235ccbe59c4964526) when we
migrated wfGetDB() callers for this extension. Ever since then,
we've done some refactoring and there is now a central class for
getting a DB connection, specifically Utils::getCentralDB( ... )

This class property is now unused and the patch removes it.

Change-Id: Iff91ddb88fc1a6640ea06ba09011089699891995